@@ -15,31 +15,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.1.1",
+    "@angular/animations": "14.0.0",
     "@angular/cdk": "13.1.1",
-    "@angular/common": "13.1.1",
-    "@angular/compiler": "13.1.1",
-    "@angular/core": "13.1.1",
-    "@angular/forms": "13.1.1",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/forms": "14.0.0",
     "@angular/material": "13.1.1",
-    "@angular/platform-browser": "13.1.1",
-    "@angular/platform-browser-dynamic": "13.1.1",
-    "@angular/router": "13.1.1",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
     "core-js": "3.6.5",
     "rxjs": "7.5.1",
     "tslib": "2.3.1",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.1.2",
+    "@angular-devkit/build-angular": "14.0.0",
     "@angular-eslint/builder": "13.0.1",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
     "@angular-eslint/schematics": "13.0.1",
     "@angular-eslint/template-parser": "13.0.1",
     "@angular/cli": "^15.0.0",
-    "@angular/compiler-cli": "13.1.1",
-    "@angular/language-service": "13.1.1",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
     "@cypress/schematic": "^2.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "2.0.8",
@@ -61,6 +61,6 @@
     "prettier": "2.1.1",
     "protractor": "7.0.0",
     "ts-node": "9.0.0",
-    "typescript": "~4.5.4"
+    "typescript": "~4.7.4"
   }
 }
